restructred provider, removed redundant test
