+++ src/components/guided-tour/guided-tour.constants.ts
@@ -1,0 +1,1 @@
+export const COMPLETED_TOURS_STORAGE_KEY = 'guided-tour:completed';
